requirements.txt
asgiref==3.7.2
blinker==1.7.0
channels==4.0.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.7
Django==4.2.7
django-ckeditor==6.7.0
django-js-asset==2.2.0
filelock==3.12.4
fonttools==4.48.1
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
numpy==1.26.4
packaging==23.2
pillow==10.2.0
platformdirs==3.11.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
six==1.16.0
sqlparse==0.4.4
sweetify==2.3.1
tzdata==2023.3
virtualenv==20.24.6
Werkzeug==3.0.2
whitenoise==6.6.0
